Rewrote dataloading code for better clarity
